requirements.txt
Django==4.*
djangorestframework==3.14.*
jsonschema==4.19.*
pyutils @ git+https://github.com/iagocanalejas/pyutils.git@master
